Reorganize CharacterControllerPass into standalone functions and modules
