+++ common/src/character_controller/collision.rs
@@ -1,0 +1,108 @@
+//! This module is used to encapsulate character collision checking for the character controller
+
+use tracing::error;
+
+use crate::{
+    graph_collision, math,
+    node::{ChunkLayout, DualGraph},
+    proto::Position,
+};
+
+/// Checks for collisions when a character moves with a character-relative displacement vector of `relative_displacement`.
+pub fn check_collision(
+    collision_context: &CollisionContext,
+    position: &Position,
+    relative_displacement: &na::Vector3<f32>,
+) -> CollisionCheckingResult {
+    // Split relative_displacement into its norm and a unit vector
+    let relative_displacement = relative_displacement.to_homogeneous();
+    let displacement_sqr = relative_displacement.norm_squared();
+    if displacement_sqr < 1e-16 {
+        // Fallback for if the displacement vector isn't large enough to reliably be normalized.
+        // Any value that is sufficiently large compared to f32::MIN_POSITIVE should work as the cutoff.
+        return CollisionCheckingResult::stationary();
+    }
+
+    let displacement_norm = displacement_sqr.sqrt();
+    let displacement_normalized = relative_displacement / displacement_norm;
+
+    let ray = graph_collision::Ray::new(math::origin(), displacement_normalized);
+    let tanh_distance = displacement_norm.tanh();
+
+    let cast_hit = graph_collision::sphere_cast(
+        collision_context.radius,
+        collision_context.graph,
+        &collision_context.chunk_layout,
+        position,
+        &ray,
+        tanh_distance,
+    );
+
+    let cast_hit = match cast_hit {
+        Ok(r) => r,
+        Err(e) => {
+            error!("Collision checking returned {:?}", e);
+            return CollisionCheckingResult::stationary();
+        }
+    };
+
+    let distance = cast_hit
+        .as_ref()
+        .map_or(tanh_distance, |hit| hit.tanh_distance)
+        .atanh();
+
+    let displacement_vector = displacement_normalized.xyz() * distance;
+    let displacement_transform = math::translate_along(&displacement_vector);
+
+    CollisionCheckingResult {
+        displacement_vector,
+        displacement_transform,
+        collision: cast_hit.map(|hit| Collision {
+            // `CastEndpoint` has its `normal` given relative to the character's original position,
+            // but we want the normal relative to the character after the character moves to meet the wall.
+            // This normal now represents a contact point at the origin, so we omit the w-coordinate
+            // to ensure that it's orthogonal to the origin.
+            normal: na::UnitVector3::new_normalize(
+                (math::mtranspose(&displacement_transform) * hit.normal).xyz(),
+            ),
+        }),
+    }
+}
+
+/// Contains information about the character and the world that is only relevant for collision checking
+pub struct CollisionContext<'a> {
+    pub graph: &'a DualGraph,
+    pub chunk_layout: ChunkLayout,
+    pub radius: f32,
+}
+
+pub struct CollisionCheckingResult {
+    /// The displacement allowed for the character before hitting a wall. The result of
+    /// `math::translate_along(&displacement_vector)` is `displacement_transform`.
+    pub displacement_vector: na::Vector3<f32>,
+
+    /// Multiplying the character's position by this matrix will move the character as far as it can up to its intended
+    /// displacement until it hits the wall.
+    pub displacement_transform: na::Matrix4<f32>,
+
+    pub collision: Option<Collision>,
+}
+
+impl CollisionCheckingResult {
+    /// Return a CollisionCheckingResult with no movement and no collision; useful if the character is not moving
+    /// and has nothing to check collision against. Also useful as a last resort fallback if an unexpected error occurs.
+    pub fn stationary() -> CollisionCheckingResult {
+        CollisionCheckingResult {
+            displacement_vector: na::Vector3::zeros(),
+            displacement_transform: na::Matrix4::identity(),
+            collision: None,
+        }
+    }
+}
+
+pub struct Collision {
+    /// This collision normal faces away from the collision surface and is given in the perspective of the character
+    /// _after_ it is transformed by `allowed_displacement`. The 4th coordinate of this normal vector is assumed to be
+    /// 0.0 and is therefore omitted.
+    pub normal: na::UnitVector3<f32>,
+}
